:memo: move files signup files com.alt.letseatingtime -> com.alt.letseatingtime_android.ui.fragment.login
