+++ app/src/main/java/com/alt/letseatingtime_android/ui/fragment/signup/SignupFragment2.kt
@@ -1,4 +1,4 @@
-package com.alt.letseatingtime_android.ui.fragment
+package com.alt.letseatingtime_android.ui.fragment.signup
 
 import androidx.fragment.app.Fragment
 import android.os.Bundle
